createPIDinfo moved to logtools, Fixes for tests (MainController constructor changes)
